+++ bench-vortex/src/bin/clickbench.rs
@@ -336,20 +336,17 @@ fn init_data_source(
                 if base_url.scheme() == "file" {
                     // Use the dataset to register tables.
                     let dataset = bench_vortex::BenchmarkDataset::ClickBench { single_file };
-                    // We need to keep this map_err since () doesn't implement into anyhow::Error
                     let file_path = base_url
                         .to_file_path()
                         .map_err(|_| anyhow::anyhow!("invalid file URL: {}", base_url))?;
 
-                    bench_vortex::data_manager::convert_parquet_to_vortex(&file_path, dataset)
-                        .await?;
+                    bench_vortex::file::convert_parquet_to_vortex(&file_path, dataset).await?;
                 }
 
                 match engine_ctx {
                     EngineCtx::DataFusion(ctx) => {
                         // Use the dataset to register tables.
                         let dataset = bench_vortex::BenchmarkDataset::ClickBench { single_file };
-                        // Using anyhow's ? operator will automatically capture the error
                         dataset.register_tables(&ctx.session, base_url, file_format)?;
                     }
