+++ bench-vortex/src/datasets/mod.rs
@@ -7,7 +7,7 @@ use url::Url;
 use vortex::ArrayRef;
 
 pub mod data_downloads;
-pub mod data_manager;
+pub mod file;
 pub mod struct_list_of_ints;
 pub mod taxi_data;
 pub mod tpch_l_comment;
