+++ bench-vortex/src/tpch/execute.rs
@@ -1,7 +1,7 @@
 use std::sync::Arc;
 
-use datafusion::prelude::SessionContext;
 use datafusion::physical_plan::ExecutionPlan;
+use datafusion::prelude::SessionContext;
 
 use crate::engines::df::execute_query;
